+++ attributes/default.rb
@@ -12,3 +12,8 @@ default['jlsolrcloud']['install']['version'] = '5.0.0'
 
 # Set this to false to NOT use logrotate to mange solr log
 default['jlsolrcloud']['log_rotate'] = true
+
+# Set this to true to enable remote JMX RMI connector applications
+default['jlsolrcloud']['remote_jmx'] = false
+# Set RMI port for use by JMX below if enabling remote_jmx
+# default['jlsolrcloud']['rmi_port'] = '18985'